More setup and tuning

+++ typeclasses/corpse.py
@@ -1,4 +1,3 @@
-
 import random
 
 from evennia import CmdSet, Command, DefaultObject
@@ -18,14 +17,17 @@ class Corpse(DefaultObject):
         """
         self.db.desc = "a corpse"
         self.db.power = 0
-        
+
         delay(60, self.decay)
-    
+
     def decay(self):
         """
         This function is called when the corpse decays.
 
         """
-        self.location.msg_contents(f"The {self.key} decays and disappears.", from_obj=self)
-        self.delete()
+        if self:
+            self.location.msg_contents(
+                f"The {self.key} decays and disappears.", from_obj=self
+            )
+            self.delete()
         return
